requirements.txt
gitdb==4.0.5
GitPython==3.1.12
PyYAML==5.4.1
smmap==3.0.5
